Move `StubStore into it's own module.
