+++ test/StoreitValueTest.js
@@ -1,45 +1,6 @@
 /*jshint expr: true */
-import _ from "underscore";
-import { makeEmitter } from "pubit-as-promised";
 import StoreitValue from "../src/StoreitValue";
-
-function StubStore() {
-    var data = Object.create(null);
-    var publish = makeEmitter(this, ["modified"]);
-
-    data["2"] = { color: "gray", message: "prepopulated" };
-
-    return _.extend(this, {
-        options: {
-            primaryKey: "id"
-        },
-        has: sinon.spy((id) => {
-            return id in data;
-        }),
-        get: sinon.spy((id) => {
-            if (!this.has(id)) {
-                throw new Error("Key does not exist.");
-            }
-            return data[id];
-        }),
-        set: sinon.spy((...args) => {
-            var [id, val] = args;
-
-            if (args.length === 1) {
-                val = id;
-                id = val.id;
-            }
-
-            if (this.has(id)) {
-                _.extend(this.get(id), val);
-            } else {
-                data[id] = val;
-            }
-
-            publish("modified", val, id);
-        })
-    });
-}
+import StubStore from "./support/StubStore";
 
 describe("StoreitValue", function () {
     beforeEach(() => {
